Refactor Haplotag Structure and Introduce Logging and Parsing Enhancements
- Remove `HaplotagBase.h` and update includes to utilize `HaplotagType.h` and `HaplotagParsingBam.h`, streamlining the codebase.
- Introduce `HaplotagLogging` functionality to manage logging of read HP distribution and coverage regions, enhancing traceability of results.
- Implement `HaplotagVcfParser` for improved variant file parsing, supporting both compressed and uncompressed VCF formats.
- Update `HaplotagProcess` to reflect changes in header inclusions and enhance overall structure for better maintainability.
- Modify `Makefile` to include new source files for proper compilation of logging and parsing functionalities.
