+++ Haplotag.cpp
@@ -1,6 +1,5 @@
 #include "Haplotag.h"
 #include "HaplotagProcess.h"
-#include "SomaticHaplotagProcess.h"
 #include "Util.h"
 #include <getopt.h>
