+++ HaplotagParsingBam.cpp
@@ -1,4 +1,4 @@
-#include "HaplotagBase.h"
+#include "HaplotagParsingBam.h"
 
 BamFileRAII::BamFileRAII(
     const std::string& BamFile
@@ -111,6 +111,7 @@ void BamFileRAII::destroy(){
     if (out) sam_close(out);
     isReleased = true;
 }
+
 
 HaplotagBamParser::HaplotagBamParser(
     ParsingBamMode mode,
@@ -1107,702 +1108,4 @@ int SomaticJudgeBase::determineReadHP(std::map<int, int> &hpCount, int &pqValue,
     }
 
     return hpResult;
-}
-
-void chrReadHpResult::recordReadHp(int &pos, int &hpResult, int &BaseHP){
-    posReadHpResult[pos].readHpCounter[hpResult]++;
-    
-    if(hpResult != ReadHP::unTag){
-        if(BaseHP == SnpHP::SOMATIC_H3){
-            if(hpResult != ReadHP::H1_1 && hpResult != ReadHP::H2_1 && hpResult != ReadHP::H3){
-                std::cerr << "Error(recordReadHp) => error read hp : BaseHP: " <<BaseHP << " readHP: " << hpResult << " pos: " << pos+1 << std::endl; 
-                exit(1);
-            }
-            posReadHpResult[pos].somaticSnpH3count++;
-            posReadHpResult[pos].somaticBaseReadHpCounter[hpResult]++;
-        }
-    }
-}
-
-void chrReadHpResult::recordDeriveHp(int &pos, int &deriveHP, float deriveHPsimilarity){
-    if(deriveHP != SnpHP::GERMLINE_H1 && deriveHP != SnpHP::GERMLINE_H2 && deriveHP != SnpHP::NONE_SNP){
-        std::cerr << "Error(recordDeriveHp) => error derive hp : pos: " <<pos+1 << " deriveHP: " << deriveHP << std::endl; 
-        exit(1);        
-    }
-    posReadHpResult[pos].deriveHP = deriveHP;
-    if(deriveHPsimilarity != 0.0){
-        posReadHpResult[pos].deriveHPsimilarVec.emplace_back(deriveHPsimilarity);
-        if(deriveHPsimilarity != 1.0){
-            // std::cout << "deriveHPsimilarity: " << deriveHPsimilarity << "\n";
-            // std::cout << "deriveHPsimilarityVec: " << varReadHpResult[pos].deriveHPsimilarVec.back() << "\n";
-        }
-    }
-}
-
-void chrReadHpResult::recordAlignCoverRegion(int& curVarPos, int &startPos, int &endPos){
-    if(posReadHpResult[curVarPos].coverRegionStartPos > startPos){
-        posReadHpResult[curVarPos].coverRegionStartPos = startPos;
-    }
-    if(posReadHpResult[curVarPos].coverRegionEndPos < endPos){
-        posReadHpResult[curVarPos].coverRegionEndPos = endPos;
-    }
-}
-
-
-ReadHpDistriLog::ReadHpDistriLog(){
-
-}
-
-ReadHpDistriLog::~ReadHpDistriLog(){
-
-}
-
-void ReadHpDistriLog::loadChrKey(const std::string &chr){
-    chrVarReadHpResult[chr] = chrReadHpResult();
-}
-
-chrReadHpResult* ReadHpDistriLog::getChrHpResultsPtr (const std::string &chr){
-    return &(chrVarReadHpResult[chr]);
-}
-
-void ReadHpDistriLog::recordChrReadHp(const std::string &chr, int &pos, int &hpResult, int &BaseHP){
-    //only use in single thread scenario
-    chrVarReadHpResult[chr].recordReadHp(pos, hpResult, BaseHP);
-}
-
-void ReadHpDistriLog::recordChrDeriveHp(const std::string &chr, int &pos, int &deriveHP, float deriveHPsimilarity){
-    //only use in single thread scenario
-    chrVarReadHpResult[chr].recordDeriveHp(pos, deriveHP, deriveHPsimilarity);
-}
-
-void ReadHpDistriLog::recordChrAlignCoverRegion(const std::string &chr, int &pos, int &startPos, int &endPos){
-    //only use in single thread scenario
-    chrVarReadHpResult[chr].recordAlignCoverRegion(pos, startPos, endPos);
-}
-
-
-void ReadHpDistriLog::writeReadHpDistriLog(const HaplotagParameters &params, std::string logPosfix, const std::vector<std::string> &chrVec){
-    std::ofstream *readHpDistriLog=NULL;
-    readHpDistriLog=new std::ofstream(params.resultPrefix + logPosfix);
-
-    int somaticSnpCount = 0;
-    for(auto chr: chrVec){
-        if(!chrVarReadHpResult[chr].posReadHpResult.empty()){
-            somaticSnpCount += chrVarReadHpResult[chr].posReadHpResult.size();
-        }
-    }
-
-    if(!readHpDistriLog->is_open()){
-        std::cerr<< "Fail to open write file: " << params.resultPrefix + logPosfix << "\n";
-        exit(1);
-    }else{
-        (*readHpDistriLog) << "####################################\n";
-        (*readHpDistriLog) << "# Somatic SNP read HP distribution #\n";
-        (*readHpDistriLog) << "####################################\n";
-        (*readHpDistriLog) << "##MappingQualityThreshold:"        << params.qualityThreshold << "\n";
-        (*readHpDistriLog) << "##SomaticSNP: " << somaticSnpCount << "\n";
-        (*readHpDistriLog) << "#Chr\t"
-                            << "Pos\t"
-                            << "DeriveHP\t"
-                            << "DeriveHPsimilarity\t\t"
-                            << "AltCount\t"
-                            << "somaticBase_H1-1\t"
-                            << "somaticBase_H2-1\t"
-                            << "somaticBase_H3\t\t"
-                            << "HP1read\t"
-                            << "HP2read\t"
-                            << "HP1-1read\t"
-                            << "HP2-1read\t"
-                            << "HP3read\t"
-                            << "untagRead\t"
-                            << "HP1ratio\t"
-                            << "HP2ratio\t"
-                            << "HP1-1ratio\t"
-                            << "HP2-1ratio\t"
-                            << "HP3ratio\n";
-    }
-
-    for(auto chr: chrVec){
-        std::map<int, ReadHpResult>::iterator curVarReadHpIter = chrVarReadHpResult[chr].posReadHpResult.begin();
-        while(curVarReadHpIter != chrVarReadHpResult[chr].posReadHpResult.end()){
-            int pos = (*curVarReadHpIter).first + 1;
-            int HP1readCount = (*curVarReadHpIter).second.readHpCounter[ReadHP::H1];
-            int HP1_1readCount = (*curVarReadHpIter).second.readHpCounter[ReadHP::H1_1];
-
-            int HP2readCount = (*curVarReadHpIter).second.readHpCounter[ReadHP::H2];
-            int HP2_1readCount = (*curVarReadHpIter).second.readHpCounter[ReadHP::H2_1];
-
-            int HP3readCount = (*curVarReadHpIter).second.readHpCounter[ReadHP::H3];
-
-            int totaltagRead = HP1readCount + HP2readCount + HP3readCount + HP1_1readCount + HP2_1readCount;
-
-            float HP1readRatio = (float)HP1readCount / (float)totaltagRead; 
-            float HP1_1readRatio = (float)HP1_1readCount / (float)totaltagRead; 
-
-            float HP2readRatio = (float)HP2readCount / (float)totaltagRead; 
-            float HP2_1readRatio = (float)HP2_1readCount / (float)totaltagRead; 
-
-            float HP3readRatio = (float)HP3readCount / (float)totaltagRead; 
-
-            float meanDeriveHPsimilarity = 0.0;
-            if((*curVarReadHpIter).second.deriveHPsimilarVec.size() != 0){
-                float size = (*curVarReadHpIter).second.deriveHPsimilarVec.size();
-                for(float deriveHPsimilarity: (*curVarReadHpIter).second.deriveHPsimilarVec){
-                    meanDeriveHPsimilarity += deriveHPsimilarity;
-                    // std::cout << deriveHPsimilarity << "\t";
-                }
-                meanDeriveHPsimilarity /= size;
-            }
-            (*readHpDistriLog) << std::fixed << std::setprecision(3) 
-                                << chr << "\t"
-                                << pos << "\t"
-                                << "H" << (*curVarReadHpIter).second.deriveHP << "\t"
-                                << meanDeriveHPsimilarity << "\t\t"
-                                << (*curVarReadHpIter).second.somaticSnpH3count << "\t"
-                                << (*curVarReadHpIter).second.somaticBaseReadHpCounter[ReadHP::H1_1] << "\t"
-                                << (*curVarReadHpIter).second.somaticBaseReadHpCounter[ReadHP::H2_1] << "\t"
-                                << (*curVarReadHpIter).second.somaticBaseReadHpCounter[ReadHP::H3] << "\t\t"
-                                << HP1readCount << "\t"
-                                << HP2readCount << "\t\t"
-                                << HP1_1readCount << "\t"
-                                << HP2_1readCount << "\t"
-                                << HP3readCount << "\t"
-                                << (*curVarReadHpIter).second.readHpCounter[ReadHP::unTag] << "\t"
-                                << HP1readRatio << "\t"
-                                << HP2readRatio << "\t"
-                                << HP1_1readRatio << "\t"
-                                << HP2_1readRatio << "\t"
-                                << HP3readRatio << "\n";
-            curVarReadHpIter++;
-        }
-    }
-    (*readHpDistriLog).close();
-    delete readHpDistriLog;
-    readHpDistriLog = nullptr;
-}
-
-void ReadHpDistriLog::writePosCoverRegionLog(const HaplotagParameters &params, std::string logPosfix, const std::vector<std::string> &chrVec){
-    std::ofstream *posCoverRegionLog=NULL;
-    posCoverRegionLog=new std::ofstream(params.resultPrefix + logPosfix);
-
-    int somaticSnpCount = 0;
-    for(auto chr: chrVec){
-        if(!chrVarReadHpResult[chr].posReadHpResult.empty()){
-            somaticSnpCount += chrVarReadHpResult[chr].posReadHpResult.size();
-        }
-    }
-
-    if(!posCoverRegionLog->is_open()){
-        std::cerr<< "Fail to open write file: " << params.resultPrefix + logPosfix << "\n";
-        exit(1);
-    }else{
-        (*posCoverRegionLog) << "############################\n";
-        (*posCoverRegionLog) << "# Somatic SNP cover region #\n";
-        (*posCoverRegionLog) << "############################\n";
-        (*posCoverRegionLog) << "##MappingQualityThreshold:"        << params.qualityThreshold << "\n";
-        (*posCoverRegionLog) << "##SomaticSNP: " << somaticSnpCount << "\n";
-        (*posCoverRegionLog) << "#Chr\t"
-                              << "Pos\t"
-                              << "Type\t"
-                              << "StartPos\t"
-                              << "EndPos\n";
-    }
-
-    for(auto chr: chrVec){
-        std::map<int, ReadHpResult>::iterator curVarReadHpIter = chrVarReadHpResult[chr].posReadHpResult.begin();
-        while(curVarReadHpIter != chrVarReadHpResult[chr].posReadHpResult.end()){
-            int pos = (*curVarReadHpIter).first + 1;
-
-            (*posCoverRegionLog) << std::fixed << std::setprecision(3) 
-                                << chr << "\t"
-                                << pos << "\t"
-                                << "somatic" << "\t"
-                                << (*curVarReadHpIter).second.coverRegionStartPos << "\t"
-                                << (*curVarReadHpIter).second.coverRegionEndPos << "\n";
-            curVarReadHpIter++;
-        }
-    }
-    (*posCoverRegionLog).close();
-    delete posCoverRegionLog;
-    posCoverRegionLog = nullptr;
-}
-
-void ReadHpDistriLog::writeTagReadCoverRegionLog(const HaplotagParameters &params, std::string logPosfix, const std::vector<std::string> &chrVec, std::map<std::string, int> &chrLength){
-    std::ofstream *tagReadCoverRegionLog=NULL;
-    tagReadCoverRegionLog=new std::ofstream(params.resultPrefix + logPosfix);
-
-    std::map<std::string, std::vector<coverRegionInfo>> coverRegion;
-
-    // merge cover region from different SNPs
-    for (const auto& chr : chrVec){
-        auto curVarReadHpIter = chrVarReadHpResult[chr].posReadHpResult.begin();
-
-        if (curVarReadHpIter == chrVarReadHpResult[chr].posReadHpResult.end()) {
-            continue;
-        }
-
-        int curStartPos = curVarReadHpIter->second.coverRegionStartPos;
-        int curEndPos = curVarReadHpIter->second.coverRegionEndPos;
-
-        while (curVarReadHpIter != chrVarReadHpResult[chr].posReadHpResult.end()){
-
-            auto nextVarReadHpIter = std::next(curVarReadHpIter);
-            if(nextVarReadHpIter != chrVarReadHpResult[chr].posReadHpResult.end()){
-                int nextStartPos = nextVarReadHpIter->second.coverRegionStartPos;
-                int nextEndPos = nextVarReadHpIter->second.coverRegionEndPos;
-
-                // region not overlap
-                if (curEndPos < nextStartPos) {
-                    coverRegionInfo regionInfo;
-                    regionInfo.startPos = curStartPos;
-                    regionInfo.endPos = curEndPos;
-                    regionInfo.length = curEndPos - curStartPos + 1;
-                    coverRegion[chr].emplace_back(regionInfo);
-                    curStartPos = nextStartPos;
-                    curEndPos = nextEndPos;
-                }else{
-                    // regions overlap
-                    curStartPos = std::min(curStartPos, nextStartPos);
-                    curEndPos = std::max(curEndPos, nextEndPos);
-                }
-            // last region
-            }else{
-                coverRegionInfo regionInfo;
-                regionInfo.startPos = curStartPos;
-                regionInfo.endPos = curEndPos;
-                regionInfo.length = curEndPos - curStartPos + 1;
-                coverRegion[chr].emplace_back(regionInfo);
-                break;
-            }
-
-            curVarReadHpIter++;
-        }
-    }
-
-    std::map<std::string, float> coverRegionRatio;
-    long long totalChrLength = 0;
-    long long totalChrCoverLength = 0;
-    double totalChrCoverageRatio = 0.0;
-
-    // calculate cover region ratio
-    for(auto chr: chrVec){
-        int totalCoverLength = 0;
-        auto coverRegionIter = coverRegion[chr].begin();
-        while(coverRegionIter != coverRegion[chr].end()){
-            totalCoverLength += (*coverRegionIter).length;
-            coverRegionIter++;
-        }
-        coverRegionRatio[chr] = (float)totalCoverLength / (float)chrLength[chr];
-        totalChrLength += chrLength[chr];
-        totalChrCoverLength += totalCoverLength;
-    }
-    totalChrCoverageRatio = (double)totalChrCoverLength / (double)totalChrLength;
-
-    if(!tagReadCoverRegionLog->is_open()){
-        std::cerr<< "Fail to open write file: " << params.resultPrefix + logPosfix << "\n";
-        exit(1);
-    }else{
-        (*tagReadCoverRegionLog) << "##################################\n";
-        (*tagReadCoverRegionLog) << "# Somatic reads cover region bed #\n";
-        (*tagReadCoverRegionLog) << "##################################\n";
-        (*tagReadCoverRegionLog) << "##MappingQualityThreshold: "   << params.qualityThreshold << "\n";
-        (*tagReadCoverRegionLog) << "##----Chr coverage ratio----\n";
-        (*tagReadCoverRegionLog) << "##Total chr coverage ratio: " << totalChrCoverageRatio << "\n";
-        for(auto chr: chrVec){
-            (*tagReadCoverRegionLog) <<"##" << chr << ":" << coverRegionRatio[chr] << "\n";
-        }
-        (*tagReadCoverRegionLog) << "#Chr\t"
-                                 << "StartPos\t"
-                                 << "EndPos\n";
-                                // << "length\n";
-    }
-
-    for(auto chr: chrVec){
-        auto coverRegionIter = coverRegion[chr].begin();
-        while(coverRegionIter != coverRegion[chr].end()){
-
-            (*tagReadCoverRegionLog) << std::fixed << std::setprecision(3) 
-                                     << chr << "\t"
-                                     << (*coverRegionIter).startPos << "\t"
-                                     << (*coverRegionIter).endPos << "\n";
-                                     //<< (*coverRegionIter).length << "\n";
-            coverRegionIter++;
-        }
-    }
-    (*tagReadCoverRegionLog).close();
-    delete tagReadCoverRegionLog;
-    tagReadCoverRegionLog = nullptr;
-}
-
-
-void ReadHpDistriLog::removeNotDeriveByH1andH2pos(const std::vector<std::string> &chrVec){
-    for (const auto& chr : chrVec) {
-        auto& chrResult = chrVarReadHpResult[chr];
-        for (auto it = chrResult.posReadHpResult.begin(); it != chrResult.posReadHpResult.end(); ) {
-            if (!it->second.existDeriveByH1andH2) {
-                //std::cerr << "Removed position not derived by H1 and H2: " << chr << " " << it->first << std::endl;
-                it = chrResult.posReadHpResult.erase(it);
-            } else {
-                ++it;
-            }
-        }
-    }
-}
-
-
-VcfParser::VcfParser(Genome tagGeneType){
-    this->tagGeneType = tagGeneType;
-    reset();
-}
-
-VcfParser::VcfParser(){
-    reset();
-}
-
-VcfParser::~VcfParser(){
-
-}
-
-void VcfParser::reset(){
-    parseSnpFile = false;
-    parseSVFile = false;
-    parseMODFile = false;
-    integerPS = false;
-    psIndex.clear();
-}
-
-void VcfParser::variantParser(std::string &variantFile, VCF_Info &Info, std::map<std::string, std::map<int, MultiGenomeVar>> &mergedChrVarinat){
-
-    if( variantFile.find("gz") != std::string::npos ){
-        // .vcf.gz 
-        compressParser(variantFile, Info, mergedChrVarinat);
-    }
-    else if( variantFile.find("vcf") != std::string::npos ){
-        // .vcf
-        unCompressParser(variantFile, Info, mergedChrVarinat);
-    }
-    else{
-        std::cerr<<"file: "<< variantFile << "\nnot vcf file. please check filename extension\n";
-        exit(EXIT_FAILURE);
-    }
-    return;
-}
-
-void VcfParser::compressParser(std::string &variantFile, VCF_Info &Info, std::map<std::string, std::map<int, MultiGenomeVar>> &mergedChrVarinat){
-    gzFile file = gzopen(variantFile.c_str(), "rb");
-    if(variantFile=="")
-        return;
-    if(!file){
-        std::cerr<< "Fail to open vcf: " << variantFile << "\n";
-    }
-    else{  
-        int buffer_size = 1048576; // 1M
-        char* buffer = (char*) malloc(buffer_size);
-        if(!buffer){
-            std::cerr<<"Failed to allocate buffer\n";
-            exit(EXIT_FAILURE);
-        }
-        char* offset = buffer;
-            
-        while(true) {
-            int len = buffer_size - (offset - buffer);
-            if (len == 0){
-                buffer_size *= 2; // Double the buffer size
-                char* new_buffer = (char*) realloc(buffer, buffer_size);
-                if(!new_buffer){
-                    std::cerr<<"Failed to allocate buffer\n";
-                    free(buffer);
-                    exit(EXIT_FAILURE);
-                }
-                buffer = new_buffer;
-                offset = buffer + buffer_size / 2; // Update the offset pointer to the end of the old buffer
-                len = buffer_size - (offset - buffer);
-            }
-
-            len = gzread(file, offset, len);
-            if (len == 0) break;    
-            if (len <  0){ 
-                int err;
-                fprintf (stderr, "Error: %s.\n", gzerror(file, &err));
-                exit(EXIT_FAILURE);
-            }
-
-            char* cur = buffer;
-            char* end = offset+len;
-            for (char* eol; (cur<end) && (eol = std::find(cur, end, '\n')) < end; cur = eol + 1)
-            {
-                std::string input = std::string(cur, eol);
-                parserProcess(input, Info, mergedChrVarinat);
-            }
-            // any trailing data in [eol, end) now is a partial line
-            offset = std::copy(cur, end, buffer);
-        }
-        gzclose (file);
-        free(buffer);
-    }    
-}
-
-void VcfParser::unCompressParser(std::string &variantFile, VCF_Info &Info, std::map<std::string, std::map<int, MultiGenomeVar>> &mergedChrVarinat){
-    std::ifstream originVcf(variantFile);
-    if(variantFile=="")
-        return;
-    if(!originVcf.is_open()){
-        std::cerr<< "Fail to open vcf: " << variantFile << "\n";
-        exit(1);
-    }
-    else{
-        std::string input;
-        while(! originVcf.eof() ){
-            std::getline(originVcf, input);
-            parserProcess(input, Info, mergedChrVarinat);
-        }
-    }
-}
-
-void VcfParser::setParseSnpFile(bool parseSnpFile){
-    this->parseSnpFile = parseSnpFile;
-}
-
-void VcfParser::setParseSVFile(bool parseSVFile){
-    this->parseSVFile = parseSVFile;
-} 
-
-void VcfParser::setParseMODFile(bool parseMODFile){
-    this->parseMODFile = parseMODFile;
-}
-
-bool VcfParser::getParseSnpFile(){
-    return this->parseSnpFile;
-}
-
-void VcfParser::parserProcess(std::string &input, VCF_Info &Info, std::map<std::string, std::map<int, MultiGenomeVar>> &mergedChrVarinat){
-    if( input.substr(0, 2) == "##" && parseSnpFile){
-        if( input.find("contig=")!= std::string::npos ){
-            int id_start  = input.find("ID=")+3;
-            int id_end    = input.find(",length=");
-            int len_start = id_end+8;
-            int len_end   = input.find(">");
-            
-            std::string chr = input.substr(id_start,id_end-id_start);
-            int chrLen = std::stoi( input.substr(len_start,len_end-len_start) );
-
-            Info.chrVec.push_back(chr);
-            Info.chrLength[chr]=chrLen;                
-        }
-        if( input.substr(0, 16) == "##FORMAT=<ID=PS," ){
-            if( input.find("Type=Integer")!= std::string::npos ){
-                integerPS = true;
-            }
-            else if( input.find("Type=String")!= std::string::npos ){
-                integerPS = false;
-                std::cerr<< "PS type is String. Auto index to integer ... ";
-            }
-            else{
-                std::cerr<< "ERROR: not found PS type (Type=Integer or Type=String).\n"; 
-                exit(EXIT_SUCCESS);
-            }
-        }
-    }
-    else if ( input.substr(0, 1) == "#" ){
-        
-    }
-    else{
-        std::istringstream iss(input);
-        std::vector<std::string> fields((std::istream_iterator<std::string>(iss)),std::istream_iterator<std::string>());
-
-        if( fields.size() == 0 )
-            return;
-
-        // trans to 0-base
-        int pos = std::stoi( fields[1] ) - 1;
-        std::string chr = fields[0];
-                
-        // find GT flag colon
-        int colon_pos = 0;
-        int gt_pos = fields[8].find("GT");
-        for(int i =0 ; i< gt_pos ; i++){
-            if(fields[8][i]==':')
-                colon_pos++;
-        }
-        // find GT value start
-        int current_colon = 0;
-        int modifu_start = 0;
-        for(unsigned int i =0; i < fields[9].length() ; i++){
-            if( current_colon >= colon_pos )
-                break;
-            if(fields[9][i]==':')
-                current_colon++;  
-            modifu_start++;
-        }
-
-        // phased hetero GT
-        if( (fields[9][modifu_start] != fields[9][modifu_start+2]) && fields[9][modifu_start+1] == '|' ){
-            // find PS flag
-            colon_pos = 0;
-            int ps_pos = fields[8].find("PS");
-            for(int i =0 ; i< ps_pos ; i++){
-                if(fields[8][i]==':')
-                    colon_pos++;
-            }
-
-            // find PS value start
-            current_colon = 0;
-            int ps_start = 0;
-            for(unsigned int i =0; i < fields[9].length() ; i++){
-                if( current_colon >= colon_pos )
-                    break;
-                if(fields[9][i]==':')
-                    current_colon++;  
-                ps_start++;
-            }
-            
-            std::string psValue;
-            // get PS value
-            if( fields[9].find(":",ps_start+1) != std::string::npos ){
-                int ps_end_pos = fields[9].find(":",ps_start+1);
-                psValue = fields[9].substr(ps_start, ps_end_pos - ps_start);
-            }
-            else{
-                psValue = fields[9].substr(ps_start, fields[9].length() - ps_start);
-            }
-            
-            // snp file
-            if( parseSnpFile ){
-
-                VarData varData;
-                varData.allele.Ref = fields[3];
-                varData.allele.Alt = fields[4];
-                varData.is_phased_hetero = true;
-                varData.setVariantType();
-                
-                if(integerPS){
-                    varData.PhasedSet = std::stoi(psValue);
-                }
-                else{
-                    std::map<std::string, int>::iterator psIter = psIndex.find(psValue);
-                    
-                    if( psIter == psIndex.end() ){
-                        psIndex[psValue] = psIndex.size();
-                    }
-                    varData.PhasedSet = psIndex[psValue];
-                }
-                
-                // record haplotype allele
-                if( fields[9][modifu_start] == '0' && fields[9][modifu_start+2] == '1' ){
-                    varData.HP1 = fields[3];
-                    varData.HP2 = fields[4];
-                }
-                else if( fields[9][modifu_start] == '1' && fields[9][modifu_start+2] == '0' ){
-                    varData.HP1 = fields[4];
-                    varData.HP2 = fields[3];
-                }
-
-                if(Info.gene_type == NORMAL){
-                    mergedChrVarinat[chr][pos].Variant[NORMAL] = varData;
-                }else if(Info.gene_type == TUMOR){
-                    mergedChrVarinat[chr][pos].Variant[TUMOR] = varData;
-                }
-            }
-            // sv file
-            if( parseSVFile ){
-                // get read INFO
-                int read_pos = fields[7].find("RNAMES=");
-                read_pos = fields[7].find("=",read_pos);
-                read_pos++;
-                        
-                int next_field = fields[7].find(";",read_pos);
-                std::string totalRead = fields[7].substr(read_pos,next_field-read_pos);
-                std::stringstream totalReadStream(totalRead);
-                
-                int svHaplotype;
-                // In which haplotype does SV occur
-                if( fields[9][modifu_start] == '0' && fields[9][modifu_start+2] == '1' ){
-                    svHaplotype = 1;
-                }
-                else if( fields[9][modifu_start] == '1' && fields[9][modifu_start+2] == '0' ){
-                    svHaplotype = 0;
-                }
-                
-                std::string read;
-                while(std::getline(totalReadStream, read, ','))
-                {
-                   auto readIter = Info.readSVHapCount.find(read);
-                   if(readIter==Info.readSVHapCount.end()){
-                       Info.readSVHapCount[read][0]=0;
-                       Info.readSVHapCount[read][1]=0;
-                   }
-                   Info.readSVHapCount[read][svHaplotype]++;
-                }
-                
-            }
-            // mod file
-            if( parseMODFile ){
-                // get read INFO
-                int read_pos = fields[7].find("MR=");
-                read_pos = fields[7].find("=",read_pos);
-                read_pos++;
-                        
-                int next_field = fields[7].find(";",read_pos);
-                std::string totalRead = fields[7].substr(read_pos,next_field-read_pos);
-                std::stringstream totalReadStream(totalRead);
-                
-                int modHaplotype;
-                // In which haplotype does SV occur
-                if( fields[9][modifu_start] == '0' && fields[9][modifu_start+2] == '1' ){
-                    modHaplotype = 1;
-                }
-                else if( fields[9][modifu_start] == '1' && fields[9][modifu_start+2] == '0' ){
-                    modHaplotype = 0;
-                }
-                
-                std::string read;
-                while(std::getline(totalReadStream, read, ','))
-                {
-                   auto readIter = Info.readSVHapCount.find(read);
-                   if(readIter==Info.readSVHapCount.end()){
-                       Info.readSVHapCount[read][0]=0;
-                       Info.readSVHapCount[read][1]=0;
-                   }
-                   Info.readSVHapCount[read][modHaplotype]++;
-                }
-            }
-        }
-        // record unphased tumor SNPs
-        else if((tagGeneType == Genome::TUMOR)){
-            //homozygous SNPs
-            if( fields[9][modifu_start] == '1' && fields[9][modifu_start+1] == '/' && fields[9][modifu_start+2] == '1' ){
-                if(parseSnpFile){
-
-                    VarData varData;
-                    varData.allele.Ref = fields[3];
-                    varData.allele.Alt = fields[4];
-                    varData.is_homozygous = true;
-                    varData.setVariantType();
-
-                    if(Info.gene_type == NORMAL){
-                        mergedChrVarinat[chr][pos].Variant[NORMAL] = varData;
-                    }else if(Info.gene_type == TUMOR){
-                        mergedChrVarinat[chr][pos].Variant[TUMOR] = varData;
-                    }
-                }
-            //unphased heterozygous
-            }else if( fields[9][modifu_start] == '0' && fields[9][modifu_start+1] == '/' && fields[9][modifu_start+2] == '1' ){
-                if(parseSnpFile){
-
-                    VarData varData;
-                    varData.allele.Ref = fields[3];
-                    varData.allele.Alt = fields[4];
-                    
-                    varData.is_unphased_hetero = true;
-                    varData.setVariantType();
-
-                    if(Info.gene_type == NORMAL){
-                        mergedChrVarinat[chr][pos].Variant[NORMAL] = varData;
-                    }else if(Info.gene_type == TUMOR){
-                        mergedChrVarinat[chr][pos].Variant[TUMOR] = varData;
-                    }
-                }
-            }
-        }
-    }
 }
